Refactor CSS and JS files
