+++ docs/details.3cc03f4a.js
@@ -1,0 +1,18 @@
+parcelRequire=function(e,r,t,n){var i,o="function"==typeof parcelRequire&&parcelRequire,u="function"==typeof require&&require;function f(t,n){if(!r[t]){if(!e[t]){var i="function"==typeof parcelRequire&&parcelRequire;if(!n&&i)return i(t,!0);if(o)return o(t,!0);if(u&&"string"==typeof t)return u(t);var c=new Error("Cannot find module '"+t+"'");throw c.code="MODULE_NOT_FOUND",c}p.resolve=function(r){return e[t][1][r]||r},p.cache={};var l=r[t]=new f.Module(t);e[t][0].call(l.exports,p,l,l.exports,this)}return r[t].exports;function p(e){return f(p.resolve(e))}}f.isParcelRequire=!0,f.Module=function(e){this.id=e,this.bundle=f,this.exports={}},f.modules=e,f.cache=r,f.parent=o,f.register=function(r,t){e[r]=[function(e,r){r.exports=t},{}]};for(var c=0;c<t.length;c++)try{f(t[c])}catch(e){i||(i=e)}if(t.length){var l=f(t[t.length-1]);"object"==typeof exports&&"undefined"!=typeof module?module.exports=l:"function"==typeof define&&define.amd?define(function(){return l}):n&&(this[n]=l)}if(parcelRequire=f,i)throw i;return f}({"mVQM":[function(require,module,exports) {
+module.exports={timeline:[{id:"1",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"2",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"3",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"4",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"5",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"6",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"7",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"8",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"9",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"10",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"11",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"12",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"}]};
+},{}],"VJ4p":[function(require,module,exports) {
+module.exports={details:[{id:"1",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"2",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"3",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"4",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"5",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"6",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"7",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"8",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"9",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"},{id:"10",hour:"8:00 às 10:00",local:"Jurere Internacional",gradeBold:"4.8",grade:"(1.024 avaliações)",image:"roteiro-sao-paulo.png"},{id:"11",hour:"8:00 às 15:00",local:"Rio De Janeiro",gradeBold:"4",grade:"(1.500 avaliações)",image:"roteiro-sao-paulo.png"},{id:"12",hour:"8:00 às 11:00",local:"São Paulo",gradeBold:"4.2",grade:"(1.230 avaliações)",image:"roteiro-sao-paulo.png"}]};
+},{}],"WHvu":[function(require,module,exports) {
+module.exports={modal:[{id:"1",content:"1 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"2",content:"2 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"3",content:"3 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"4",content:"4 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"5",content:"5 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"6",content:"6 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"7",content:"7 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"8",content:"8 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"9",content:"9 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"10",content:"10 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"11",content:"11 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"},{id:"12",content:"12 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",image:"destinos-gramado.png",gradeBold:"4.8",grade:"(1.024 avaliações)",local:"Jurere Internacional"}]};
+},{}],"SCVz":[function(require,module,exports) {
+"use strict";function o(){document.getElementById("mainDropdown").classList.toggle("is-visible")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.openMainDropdown=o,exports.openRoadmapDropdown=exports.openUserDropdown=void 0;var e=function(o){document.querySelector('[data-dropdown-info="'.concat(o,'"]')).classList.toggle("is-visible")};exports.openUserDropdown=e;var t=function(o){document.querySelector('[data-roadmap-info="'.concat(o,'"]')).classList.toggle("is-visible")};exports.openRoadmapDropdown=t;
+},{}],"RSqK":[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.buildModalImages=exports.buildModal=exports.openModalinfo=void 0;var a=function(a){document.querySelector('[data-modal-info="'.concat(a,'"]')).classList.toggle("is-visible")};exports.openModalinfo=a;var n=function(a){return'\n  <div class="modal" data-modal-info="'.concat(a.id,'">\n    <div class="modal__close" data-closeArea="').concat(a.id,'"></div>\n    <div class="modal__modal in-column">\n      <div class="in-row">\n        <p class="modal__title">').concat(a.local,'</p>\n        <button class="icon__close is-absolute-right" data-close="').concat(a.id,'"></button>\n      </div>\n      <hr class="underline underline--shine underline__modal--margin" />\n      <div class="details__rating details__rating--text in-row">\n        <svg class="icon__star">\n          <use xlink:href="#icon-star"></use>\n        </svg>\n        <p class="details__grade details__grade--bold">').concat(a.gradeBold,'</p>\n        <p class="details__grade">').concat(a.grade,'</p>\n        <p class="tag tag__margin">natureza</p>\n        <p class="tag tag__margin">lazer</p>\n      </div>\n      <p class="modal__text">').concat(a.content,'</p> \n      <a class="modal__link" href="#">Clique para ver a descrição completa...</a>\n      <div class="modal__gallery in-row"></div>\n      <div class="modal__footer in-row">\n        <button class="icon__arrow icon__arrow--left"></button>\n        <button class="icon__arrow icon__arrow--right"></button>\n      </div>\n    </div>\n  </div>\n  ')};exports.buildModal=n;var o=function(a){return'\n  <img class="modal__images" src="'.concat(a.image,'">\n  ')};exports.buildModalImages=o;
+},{}],"t8Xq":[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.buildTimeline=void 0;var n=function(n){return'\n    <li class="timeline__item in-column">\n      <div class="in-row timeline__time">\n        <span class="timeline__hour">'.concat(n.hour,'</span>\n        <button class="button__menu button--right button--modal" data-dropdown-click="').concat(n.id,'" onclick="openUserDropdown(').concat(n.id,')">\n          <svg class="icon__menu">\n            <use xlink:href="#icon-menu"></use>\n          </svg>\n        </button>\n        <div class="in-column">\n          <ul class="dropdown" data-dropdown-info="').concat(n.id,'">\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--details" href="#">Detalhes</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--reorder" href="#">Reordenar</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--hour" href="#">Alterar horário</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--calendar" href="#">Alterar data</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__text--red dropdown__link--delete" href="#">Remover ponto</a></li>\n          </ul>\n        </div>\n      </div>\n      <p class="timeline__local">').concat(n.local,'</p>\n      <div class="details__rating details__rating--text in-row">    \n        <svg class="icon__star">\n            <use xlink:href="#icon-star"></use>\n        </svg>\n        <p class="details__grade details__grade--bold">').concat(n.gradeBold,'</p>\n        <p class="details__grade">').concat(n.grade,'</p>\n      </div>\n      <div class="in-row timeline__category">\n        <p class="tag">natureza</p>\n        <p class="tag">lazer</p>\n      </div>\n      <hr class="underline underline--shine underline__timeline--margin" />\n    </li>\n  ')};exports.buildTimeline=n;
+},{}],"Zy0T":[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.buildDetailsList=void 0;var n=function(n){return'\n  <div class="details__items in-column">\n    <div class="details__content in-row">\n      <img class="details__image" src="'.concat(n.image,'"/>\n      <div class="details__rating--text in-column">\n          <p class="details__title">').concat(n.local,'</p>\n          <div class="details__rating in-row">\n            <svg class="icon__star">\n              <use xlink:href="#icon-star"></use>\n            </svg>\n            <p class="details__grade details__grade--bold">').concat(n.gradeBold,'</p>\n            <p class="details__grade">').concat(n.grade,'</p>\n          </div>\n          <div class="in-row">\n              <p class="tag">natureza</p>\n              <p class="tag">lazer</p>\n          </div>\n      </div>\n    </div>\n    <div class="details__options in-row">\n        <p class="details__text--options details__text--guide">3 dias</p>\n        <div class="in-column details__button--info">\n          <button class="button__menu has-margin-right" onclick="openModalinfo(').concat(n.id,')">\n            <svg class="icon__info">\n              <use xlink:href="#icon-info"></use>\n              </svg>\n            </button>\n          </div>\n        <div class="in-column details__button--menu"> \n          <button class="button__menu has-margin-right" onclick="openRoadmapDropdown(').concat(n.id,')">\n            <svg class="icon__menu">\n              <use xlink:href="#icon-menu"></use>\n            </svg>\n          </button>\n          <ul class="dropdown dropdown__option" data-roadmap-info="').concat(n.id,'">\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--check" href="#">Usar</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--save" href="#">Salvar</a></li>\n            <li class="dropdown__item"><a class="dropdown__text dropdown__link--share" href="#">Compartilhar</a></li>\n          </ul>\n        </div>\n      </div>\n    </div>\n    <hr class="underline underline--shine" />\n  ')};exports.buildDetailsList=n;
+},{}],"j6Bo":[function(require,module,exports) {
+"use strict";var e=require("./data/timeline.json"),n=require("./data/details.json"),o=require("./data/modal.json"),i=require("./js/dropdown.js"),r=require("./js/modal.js"),a=require("./js/timeline.js"),t=require("./js/list.js"),d=document.querySelector("#modal");o.modal.forEach(function(e){var n=(0,r.buildModal)(e);d.innerHTML=d.innerHTML+n});var l=document.querySelector(".modal__gallery");n.details.forEach(function(e){var n=(0,r.buildModalImages)(e);l.innerHTML=l.innerHTML+n});var s=document.querySelector(".timeline__list");e.timeline.forEach(function(e){var n=(0,a.buildTimeline)(e);s.innerHTML=s.innerHTML+n});var u=document.querySelector(".details__grid");n.details.forEach(function(e){var n=(0,t.buildDetailsList)(e);u.innerHTML=u.innerHTML+n}),document.querySelector("#mainDropdownButton").addEventListener("click",i.openMainDropdown),window.addEventListener("click",function(e){if(!e.target.matches("#mainDropdownButton")){var n=document.getElementById("mainDropdown");n.classList.contains("is-visible")&&n.classList.remove("is-visible")}}),window.openUserDropdown=i.openUserDropdown,window.openRoadmapDropdown=i.openRoadmapDropdown,window.openModalinfo=r.openModalinfo;
+},{"./data/timeline.json":"mVQM","./data/details.json":"VJ4p","./data/modal.json":"WHvu","./js/dropdown.js":"SCVz","./js/modal.js":"RSqK","./js/timeline.js":"t8Xq","./js/list.js":"Zy0T"}]},{},["j6Bo"], null)
+//# sourceMappingURL=details.3cc03f4a.js.map
