+++ src/css/components/_dropdown.scss
@@ -1,7 +1,7 @@
 .dropdown {
   background-color: white;
   border-radius: 8px;
-  box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.25);
+  box-shadow: 0 2px 2px rgba(0, 0, 0, .25);
   display: flex;
   flex-direction: column;
   margin-top: 15px;
