+++ src/css/components/_explore.scss
@@ -101,7 +101,7 @@
     margin: 0 auto;
     max-width: 80%;
     position: relative;
-    text-shadow: 0px 1px 1px rgba(0, 0, 0, 0.25);
+    text-shadow: 0 1px 1px rgba(0, 0, 0, .25);
     transition: .75s;
     z-index: 5;
   }
