+++ src/css/components/_hero.scss
@@ -19,7 +19,7 @@
     color: $white;
     font-size: 48px;
     font-weight: bold;
-    text-shadow: 0px 2px 2px rgba(0, 0, 0, 0.5);
+    text-shadow: 0 2px 2px rgba(0, 0, 0, .5);
     text-transform: uppercase;
 
     @media screen and (max-width: 440px) {
