--- src/css/components/_info.scss
@@ -1,81 +1,0 @@
-.info {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  z-index: 9999;
-  visibility: hidden;
-
-  &__close {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(10,23,55,0.2);
-  }
-
-  &__modal {
-    position: absolute;
-    display: flex;
-    border-radius: 8px;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    height: 580px;
-    margin: auto;
-    padding: 24px 2em;
-    background-color: white;
-    z-index: 9999;
-  
-    @media screen and (min-width: 820px) {
-      max-width: 800px;
-    }
-
-    @media screen and (max-width: 819px) {
-      width: 90%;
-    }
-  }
-
-  &__title {
-    color: $gray3;
-    margin: 0 0 8px;
-
-    @media screen and (min-width: 481px) {
-      font-size: 32px;
-    }
-
-    @media screen and (max-width: 480px) {
-      font-size: 28px;
-      max-width: 200px;
-    }
-  }
-
-  &__text {
-    margin-top: 25px;
-    font-size: 16px;
-    color: $gray3;
-  }
-  
-  &__link {
-    font-size: 16px;
-    color: $orange;
-    text-align: center;
-  }
-
-  &__gallery {
-    overflow-x: auto;
-    background: linear-gradient(270deg, #FFFFFF 0%, rgba(255, 255, 255, 0) 65.41%);
-  }
-
-  &__images {
-    height: 200px;
-    width: auto;
-    margin: 30px 15px 0 0;
-  }
-
-  &__footer {
-    width: 100%;
-    display: flex;
-    justify-content: flex-end;
-    align-items: flex-end;
-  }
-}
