+++ src/css/components/_timeline.scss
@@ -1,0 +1,160 @@
+.timeline {
+  background-color: $white;
+  height: 100%;
+  min-width: 300px;
+  overflow-x: hidden;
+  overflow-y: hidden;
+  padding: 20px 0 2em 20px;
+  position: relative;
+  max-width: 350px;
+
+  &__headline {
+    padding-right: 20px;
+  }
+
+  &__header {
+    width: 100%;
+    display: flex;
+    position: relative;
+    right: 5px;
+    justify-content: space-between;
+  }
+
+  &__input {
+    background-color: $white;
+    border-radius: 8px;
+    border: 1px solid $gray2;
+    display: flex;
+    height: 32px;
+    margin-top: -8px;
+    opacity: .75;
+    padding: 8px 16px;
+    width: 75%;
+  }
+
+  &__input--value {
+    color: $gray2;
+    font-size: 10px;
+  }
+
+  &__days {
+    color: $gray2;
+    font-size: 16px;
+    margin-bottom: -10px;
+    margin-left: 15px;
+  }
+
+  &__select {
+    -moz-appearance: none;
+    -webkit-appearance: none;
+    appearance: none;
+    background-color: $white;
+    background-image: url("data:image/svg+xml,%3Csvg width='10' height='8' viewBox='0 0 10 6' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M5.065 4.957.49.441a.288.288 0 0 0-.406 0 .28.28 0 0 0 0 .4L4.863 5.56a.288.288 0 0 0 .405 0L10.045.84a.274.274 0 0 0 0-.398.288.288 0 0 0-.406 0L5.065 4.957z' fill='%23444'/%3E%3C/svg%3E") ;
+    background-position-x: 23px;
+    background-position: center;
+    background-repeat: no-repeat;
+    border-radius: 8px;
+    border: 1px solid $gray2;
+    color: $black;
+    height: 22px;
+    margin: 0 8px;
+    opacity: .75;
+    padding: 0 6px;
+    width: 42px;
+  }
+
+  &__list {
+    color: $gray3;
+    counter-reset: part 1;
+    overflow-y: auto;
+    padding-right: 5px;
+    scrollbar-width: none;
+    width: 100%;
+  }
+
+  &__time {
+    position: relative;
+    display: flex;
+    justify-content:flex-end;
+    margin-top: -15px;
+    margin-bottom: 8px;
+    padding-right: 12px;
+  }
+
+  &__item:not(:first-child){
+    counter-increment: part;
+  }
+
+  &__item {
+    max-width: 280px;
+    margin-left: -10px;
+    margin-right: 10px;
+  }
+
+  &__item::before {
+    align-items: center;
+    border-radius: 25px;
+    background-color: $white;
+    font-size: 12px;
+    border: 1px solid $gray2;
+    content: "" counter(part) "";
+    display: flex;
+    height: 24px;
+    justify-content: center;
+    width: 24px;
+    position: relative;
+    left: -30px;
+    margin-top: 5px;
+    z-index: 1;
+  }
+
+  &__item:first-of-type:after {
+    position: absolute;
+    content: '';
+    width: 1px;
+    height: 100%;
+    left: 31px;
+    background-color: transparent;
+    border: 1px dashed $gray2;
+    z-index: 0;
+    margin-top: 5px;
+  }
+
+  &__item:not(:first-of-type) {
+    margin-top: 10px;
+  }
+
+  &__hour {
+    display: flex;
+    position: absolute;
+    left: 0;
+    font-size: 10px;
+    color: $gray3;
+    margin-bottom: 5px;
+    margin-top: -5px;
+  }
+
+  &__local {
+    font-size: 16px;
+    color: $gray3;
+    margin-top: -10px;
+  }
+
+  &__category {
+    margin-bottom: -5px;
+  }
+
+  &__footer {
+    align-items: center;
+    background-color: $white;
+    bottom: 0;
+    box-shadow: 0 -2px 4px rgba(0, 0, 0, .25);
+    display: flex;
+    height: 65px;
+    justify-content: center;
+    left: 0;
+    position: absolute;
+    width: 100%;
+    z-index: 999;
+  }
+}
