+++ src/css/generic/_all.scss
@@ -1,0 +1,1 @@
+@import 'normalize', 'global';
