+++ src/css/generic/_normalize.scss
@@ -110,7 +110,7 @@ dfn {
 
 h1 {
   font-size: 2em;
-  margin: 0.67em 0;
+  margin: .67em 0;
 }
 
 // Add the correct background and color in IE 9-.
@@ -304,7 +304,7 @@ textarea {
 
 ::-webkit-input-placeholder {
   color: inherit;
-  opacity: 0.54;
+  opacity: .54;
 }
 
 // 1. Correct the inability to style clickable types in iOS and Safari.
