+++ src/css/helpers/_states.scss
@@ -1,5 +1,5 @@
 .is-centered {
-  margin: 0px auto;
+  margin: 0 auto;
 }
 
 .is-relative {
