+++ src/css/main.scss
@@ -1,18 +1,17 @@
-@import 'settings/_colors.scss';
-@import 'generic/_normalize.scss';
-@import 'generic/_global.scss';
-@import 'helpers/_hovers.scss'; 
-@import 'objects/_buttons.scss';
-@import 'objects/_input.scss';
-@import 'objects/_icons.scss';
-@import 'objects/_tag.scss';
-@import 'objects/_underline.scss';
-@import 'components/_header.scss';
-@import 'components/_hero.scss';
-@import 'components/_container.scss';
-@import 'components/_explore.scss';
-@import 'components/_details.scss';
-@import 'components/_modal.scss';
-@import 'components/_dropdown.scss';
-@import 'components/_info.scss';
-@import 'helpers/_states.scss';
+// Colors, Variables, etc
+@import 'settings/_all.scss';
+
+// Mixins, Functions, etc
+@import 'tools/_all.scss'; 
+
+// Tag Styles h1, ul, button, etc
+@import 'generic/_all.scss';
+
+// Class Tag Styles .title, .list, .button, etc
+@import 'objects/_all.scss';
+
+// .header, .footer, .modal, etc
+@import 'components/_all.scss';
+
+// !important
+@import 'helpers/_all.scss';
