+++ src/css/objects/_all.scss
@@ -1,0 +1,1 @@
+@import 'buttons', 'input', 'icons', 'tag', 'underline';
