+++ src/css/objects/_buttons.scss
@@ -37,7 +37,7 @@
   &--round {
     border-radius: 50%;
     background-color: $white;
-    opacity: 0.6;
+    opacity: .6;
   }
 
   &__block {
