+++ src/css/objects/_underline.scss
@@ -13,13 +13,13 @@
     background-color: $gray1;
   }
 
-  &__info--margin {
+  &__modal--margin {
     margin-bottom: 30px;
     margin-top: 0;
   }
 
-  &__modal--margin {
-    margin-bottom: 0px;
+  &__timeline--margin {
+    margin-bottom: 0;
   }
 }
