+++ src/css/settings/_all.scss
@@ -1,0 +1,1 @@
+@import 'colors';
