--- src/data/info.json
@@ -1,100 +1,0 @@
-{
-  "info": [
-    {
-      "id": "1",
-      "content":"1 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "2",
-      "content":"2 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "3",
-      "content":"3 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image":"destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "4",
-      "content":"4 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "5",
-      "content":"5 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "6",
-      "content":"6 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "7",
-      "content":"7 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "8",
-      "content":"8 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "9",
-      "content":"9 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "10",
-      "content":"10 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "11",
-      "content":"11 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    },
-    {
-      "id": "12",
-      "content":"12 ipsum dolor sit amet, consectetur adipiscing elit. At diam in mauris varius maecenas adipiscing. Faucibus erat nisl vel, consequat aliquam vehicula neque, elementum. Non aliquam neque pharetra lacus, et elementum. Posuere ante tristique est faucibus pharetra. Turpis quisque arcu amet libero at dolor sit ipsum pellentesque.",
-      "image": "destinos-gramado.png", 
-      "gradeBold": "4.8",
-      "grade" : "(1.024 avaliações)",
-      "local": "Jurere Internacional" 
-    }
-  ]
-}
