+++ src/details/index.pug
@@ -2,9 +2,9 @@ doctype html
 html(lang="en")
   include ../templates/_head.pug
   body#details
+    include ../assets/icons.svg
     include ../templates/_header.pug
     include ../templates/_details-section.pug
-    include ../templates/_details-info.pug
-    include ../assets/icons.svg
+    include ../templates/_details-modal.pug
     script(src='../details.js' defer)
