+++ src/index.pug
@@ -2,9 +2,9 @@ doctype html
 html(lang="en")
   include ./templates/_head.pug
   body 
+    include ./assets/icons.svg
     include ./templates/_header.pug
     include ./templates/_home-hero.pug
     include ./templates/_home-explore.pug
-    include ./assets/icons.svg
     
     script(src='./index.js' defer)
