+++ src/js/dropdown.js
@@ -9,7 +9,3 @@ export const openUserDropdown = (id) => {
 export const openRoadmapDropdown = (id) => {
   var t = document.querySelector(`[data-roadmap-info="${id}"]`).classList.toggle('is-visible');
 }
-
-export const openModalinfo = (id) => {
-  document.querySelector(`[data-modal-info="${id}"]`).classList.toggle('is-visible');
-}
