--- src/js/popup.js
@@ -1,3 +1,0 @@
-export const openModalinfo = (id) => {
-  document.querySelector(`[data-modal-info="${id}"]`).classList.toggle('is-visible');
-}
