--- src/templates/_details-info.pug
@@ -1,1 +1,0 @@
-#info
