+++ src/templates/_details-section.pug
@@ -11,7 +11,7 @@ main.details
       //- Dinamic City ->
       p.details__city Florianópolis
     nav.details__nav
-      a#roteiros.details__button--nav(href="#roteiros") Roteiros
+      a#roteiros.details__button--nav.is-active(href="#roteiros") Roteiros
       a#pontosTuristicos.details__button--nav(href="#pontosTuristicos") Pontos turísticos
       hr.underline.underline--dark
       .details__grid.in-column  
@@ -28,4 +28,4 @@ main.details
         //-       svg.icon__checked
         //-         use(xlink:href="#icon-checked")
   iframe.details__map(src='https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d31753918.49138118!2d-69.64872607287624!3d-13.705681771544725!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x9c59c7ebcc28cf%3A0x295a1506f2293e63!2sBrasil!5e0!3m2!1spt-BR!2sbr!4v1621868031749!5m2!1spt-BR!2sbr' style='border:0;' allowfullscreen='' loading='lazy')
-  include ../templates/_details-modal.pug
+  include ../templates/_details-timeline.pug
