+++ src/templates/_details-timeline.pug
@@ -1,0 +1,32 @@
+section.timeline.in-column
+  header.timeline__headline
+    p.details__text--header Título do roteiro
+    input.timeline__input(placeholder="Dê um nome ao seu roteiro")
+    span.timeline__input--value 30/48
+    .in-row.timeline__header
+      .timeline__days.in-row.in-vertical-center
+        p Dia 
+        select.timeline__select(name="choice")
+          option(value="1") 1
+          option(value="2") 2
+          option(value="3") 3
+          option(value="4") 4
+          option(value="5") 5
+        p de 5
+      .in-column
+        button#mainDropdownButton.button__menu.button--margin
+          svg.icon__menu
+            use(xlink:href="#icon-menu")
+        ul#mainDropdown.dropdown.dropdown--margin
+          li.dropdown__item
+            a.dropdown__text.dropdown__link--details(href="#") Detalhes
+          li.dropdown__item
+            a.dropdown__text.dropdown__link--calendar(href="#") Editar diárias
+          li.dropdown__item
+            a.dropdown__text.dropdown__link--reorder(href="#") Reordenar
+          li.dropdown__item
+            a.dropdown__text.dropdown__text--red.dropdown__link--dump(href="#") Excluir roteiro
+    hr.underline.underline--shine
+  ul.timeline__list.in-column
+  footer.timeline__footer
+    button#save.button.button--orange.button--save Salvar
